+++ src/Strategy/Python/SetupPy.hs
@@ -1,6 +1,7 @@
 module Strategy.Python.SetupPy (
   analyze',
   installRequiresParserSetupCfg,
+  installRequiresParser,
 ) where
 
 import Control.Effect.Diagnostics
@@ -33,16 +34,29 @@ installRequiresParser = do
     Nothing -> pure []
     Just _ -> entries <* end
   where
-    prefix = skipManyTill anySingle (symbol "install_requires") *> symbol "=" *> symbol "["
-    entries = (requireSurroundedBy "\"" <|> requireSurroundedBy "\'") `sepEndBy` symbol ","
+    prefix :: Parser Text
+    prefix = skipManyTill anySingle (symbol "install_requires") *> symbol "=" *> (symbol' "[")
+
+    entries :: Parser [Req]
+    entries = entriesParser `sepEndBy` (symbol' ",")
+
+    entriesParser :: Parser Req
+    entriesParser = lexeme (requireSurroundedBy "\"" <|> requireSurroundedBy "\'")
 
     requireSurroundedBy :: Text -> Parser Req
     requireSurroundedBy quote = between (symbol quote) (symbol quote) requirementParser
 
+    end :: Parser Text
     end = symbol "]"
 
     symbol :: Text -> Parser Text
     symbol = L.symbol space
+
+    lexeme :: Parser a -> Parser a
+    lexeme = L.lexeme $ L.space space1 (L.skipLineComment "#") empty
+
+    symbol' :: Text -> Parser Text
+    symbol' = lexeme . symbol
 
 -- | Parses install requirements listed in setup.cfg
 -- Setup.cfg has install_requires attribute in [options] block